# backend/api/docs/swagger.yaml
basePath: /
definitions:
  models.CommentPost:
    properties:
      value:
        type: string
    required:
    - value
    type: object
  models.CreateOrUpdatePost:
    properties:
      message:
        minLength: 5
        type: string
      selectedFile:
        type: string
      title:
        type: string
    required:
    - message
    - title
    type: object
...
  models.PostModel:
    properties:
      _id:
        type: string
      comments:
        items:
          type: string
        type: array
      createdAt:
        type: string
      creator:
        type: string
      likes:
        items:
          type: string
        type: array
      message:
        minLength: 5
        type: string
      name:
        type: string
      selectedFile:
        type: string
      title:
        type: string
    required:
    - message
    - title
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: 这是基于 Golang Fiber 的 REST API Swagger 文档
  title: Fiber Golang Mongo Grpc Websocket 等服务
  version: "1.0"
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: 分页获取帖子列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 用户 ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PostModel'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 获取帖子列表
      tags:
      - Posts
    post:
      consumes:
      - application/json
      description: 创建一条新帖子
      parameters:
      - description: 创建帖子参数
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.CreateOrUpdatePost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建帖子
      tags:
      - Posts
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: 按帖子 ID 删除帖子，需提供创建者的认证 Token
      parameters:
      - description: 帖子 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 根据 ID 删除帖子
      tags:
      - Posts
    get:
      consumes:
      - application/json
      description: 根据帖子 ID 获取帖子详情
      parameters:
      - description: 帖子 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 获取单个帖子
      tags:
      - Posts
    patch:
      consumes:
      - application/json
      description: 更新指定 ID 的帖子
      parameters:
      - description: 帖子 ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新帖子参数
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.CreateOrUpdatePost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新帖子
      tags:
      - Posts
  /posts/{id}/commentPost:
    post:
      consumes:
      - application/json
      description: 对指定帖子新增评论
      parameters:
      - description: 帖子 ID
        in: path
        name: id
        required: true
        type: string
      - description: 评论内容
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.CommentPost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 评论帖子
      tags:
      - Posts
  /posts/{id}/likePost:
    patch:
      consumes:
      - application/json
      description: 如果当前用户未点赞则添加，已点赞则移除。
      parameters:
      - description: 帖子 ObjectID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 点赞/取消点赞帖子
      tags:
      - Posts
  /posts/search:
    get:
      consumes:
      - application/json
      description: 根据搜索关键字匹配帖子和用户
      parameters:
      - description: 搜索关键词
        in: query
        name: searchQuery
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 按关键字搜索帖子与用户
      tags:
      - Posts
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 使用 Bearer 认证，格式为 "Bearer 空格 Token"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"